Se definió el modelo de datos final. Se adjunta el DER en formato PDF, faltan atributos Se modificaron las relaciones para respetar el DER Se implementaron respuestas 409 para el borrado de una editorial o autor si poseen libros Se utilizo un hook para impedir el borrado de un autor que tiene libros (La manyToMany no permitia cambiar el schema para lograr eso). Se creo un archivo de errores y el error errorDominio para el error de borrar un autor si tiene libros. Se utiliza el error de MySQL para la respuesta al error de intentar borrar una editorial
