--- docs.md
@@ -1,23 +1,0 @@
-# Documentación
-**TODO: En revisión.**
-
-## Formato
-* La documentación debe estar disponible en un directorio **docs** en el primer nivel de el/los repositorios.
-* El punto de entrada de la misma será un archivo README.md en el directorio docs y deberá estar linkeada en el README.md del proyecto.
-* Los documentos de texto generados por los alumnos deben ser subidos en formato markdown.
-* Los documentos que sean generados directamente con una tool o app y no lo exporten en formato markdown podrán ser incluidos en formato pdf. No se admitirán archivos binarios o no formateados como .doc, .odf, .rtf, txt, etc.
-* Los diagramas, esquemas y demás contenido gráfico puede ser incluido utilizando [Mermaid](https://mermaid.js.org) o directamente en formato gráfico que sea compatible con la plataforma de git utilizada; linkeándolos en el documento correspondiente
-
-## Contenido
-A continuación se detalla la documentación mínima a incluir para cada entrega
-|Contenido - Secciones|Regularidad|Aprobación|
-|:-|:-:|:-:|
-|proposal actualizada|X|X|
-|instrucciones de instalacion|X|X|
-|links a los PR/MR|X|X|
-|minutas de reunión y avance|X|X|
-|tracking de features, bug e issues|X|X|
-|documentacion de la API||X|
-|evidencia de ejecución de test automáticos||X|
-|demo de app en video|TBD|X|
-|deploy|TBD|TBD|
