+++ src/editorial/editorial.entity.ts
@@ -6,6 +6,7 @@ import {
   Collection,
   BeforeDelete,
   EventArgs,
+  OneToMany,
 } from "@mikro-orm/core";
 
 import { BaseEntity } from "../shared/DB/baseEntity.entity.js";
@@ -17,16 +18,6 @@ export class Editorial extends BaseEntity {
   @Property()
   nombre!: string;
 
-  @ManyToMany(() => Libro, (libro) => libro.misEditoriales, {})
+  @OneToMany(() => Libro, (libro) => libro.miEditorial)
   misLibros = new Collection<Libro>(this);
-
-  @BeforeDelete()
-  async existenLibros({ em }: EventArgs<this>) {
-    const count = await em.count(Libro, { misEditoriales: this });
-    if (count > 0) {
-      throw new errorDominio(
-        "No se puede eliminar una Editorial que posea libros."
-      );
-    }
-  }
 }
