+++ src/peticionesEditorial.http
@@ -12,4 +12,4 @@ Content-Type: application/json
 }
 
 ### Delete de una editorial por ID
-DELETE  http://localhost:3000/api/editoriales/4
+DELETE  http://localhost:3000/api/editoriales/1
